Add initial Secret node
Added initial Secret node and an endpoint for requesting secrets. The endpoint requires the name of the secret and an access token that is generated upon a successful login.

In the future, if the access token is valid and the secret name is found in the database, the voting sequence will be initiated.

Addresses: #10 Create initial secret node